requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
pydantic
google-api-python-client
agno
